ensure type of published header is amqp.Header
