+++ pkg/publish.go
@@ -185,7 +185,8 @@ func (s *AmqpPublish) createWorkerFunc(
 
 				size := len((*message.Publishing).Body)
 				s.logger.Debugf("publish message to %s (%d bytes)", message.Routing, size)
-				message.Publishing.Headers = message.Routing.Headers()
+				headers := EnsureAMQPTable(message.Routing.Headers()).(amqp.Table)
+				message.Publishing.Headers = headers
 				err := session.Publish(
 					message.Routing.Exchange(),
 					message.Routing.Key(),
